fix(devshell): preserve prepare policy callsite behavior
